Add observability features with OpenTelemetry integration, including metrics, traces, and logs. Update README and documentation for setup instructions. Implement middleware for API request tracking in the Dexto agent.
